Made a frontend so you can search for recipes
Trie seems to be working, had to change the search function.

added a fake admin frontend page that allows you to manually alter ingredient names - I figure I've tweaked the parser as much as I can, will probably just filter out non fat, reduced fat, and a couple other things I noticed

certain ingredients with Capitalization are being parsed to an empty string, which makes it as if you automatically 'own' the ingredient

going to add the og tags to the db/frontend so searches look nicer

I need to figure out how to share the common folder between the modules without having to copy paste it or make it into an npm module - would a private one work in this case?